reorganize passes into into ./pipeline
